gateway/pyproject.toml
[project]
name = "narwhal-gateway"
version = "4.0.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Rajvir Jhawar", email = "rjhawar@sama.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "granian[reload]>=1.6.3",
    "httpx>=0.27.2",
    "litestar>=2.15.1",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.9.0",
    "s3fs[boto3]>=2024.12.0",
]

[dependency-groups]
dev = [
    "nodeenv>=1.9.1",
    "pyngrok>=7.2.0",
]
